tablesモジュールの切り出し、userテーブル操作をブループリントに分割
